+++ src/Xamarin.Android.Build.Tasks/Xamarin.Android.Build.Tasks.csproj
@@ -142,6 +142,7 @@
     <Compile Include="Tasks\AndroidUpdateResDir.cs" />
     <Compile Include="Tasks\AndroidSignPackage.cs" />
     <Compile Include="Tasks\RemoveDirFixed.cs" />
+    <Compile Include="Tasks\ResolveJdkJvmPath.cs" />
     <Compile Include="Tasks\ResolveSdksTask.cs" />
     <Compile Include="Tasks\CompileToDalvik.cs" />
     <Compile Include="Tasks\AndroidToolTask.cs" />
